Able to choose the port number in the java and C++ program
Java: Enter the port number in the text field then press start to create
the server socket.

C++: Then in the old program, at the bottom left corner. In the text
field below "Connect to Port", enter the same port number and press
connect.

Still need to do some error handling for both C++ and java eg, when able
to close the connection and start it again without having to close the
programs